requirements.txt
asgiref==3.5.0
azure-core==1.23.1
azure-storage-blob==12.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
pycparser==2.21
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
